Adjust speed highlighting to average speed
